Reorganize and clean up
