+++ extension/ts/background-scripts/background.js/contentScriptBergamotApiClientPortListener.ts
@@ -1,22 +1,23 @@
-import { LanguageDetector } from "./LanguageDetector";
+import { BergamotApiClient } from "./lib/BergamotApiClient";
 import { Runtime } from "webextension-polyfill-ts";
 import Port = Runtime.Port;
+const bergamotApiClient = new BergamotApiClient();
 
-export const contentScriptLanguageDetectorProxyPortListener = (port: Port) => {
-  if (port.name !== "port-from-content-script-language-detector-proxy") {
+export const contentScriptBergamotApiClientPortListener = (port: Port) => {
+  if (port.name !== "port-from-content-script-bergamot-api-client") {
     return;
   }
   port.onMessage.addListener(async function(m: {
-    str: string;
+    texts: [];
     requestId: string;
   }) {
-    // console.debug("Message from content-script-language-detector-proxy:", { m });
-    const { str, requestId } = m;
-    const results = await LanguageDetector.detectLanguage({ text: str });
-    // console.debug({ results });
+    // console.debug("Message from content-script-bergamot-api-client:", {m});
+    const { texts, requestId } = m;
+    const results = await bergamotApiClient.sendTranslationRequest(texts);
+    // console.log({ results });
     try {
       port.postMessage({
-        languageDetectorResults: {
+        translationRequestResults: {
           results,
           requestId,
         },
