+++ extension/ts/background-scripts/background.js/contentScriptLanguageDetectorProxyPortListener.ts
@@ -1,23 +1,22 @@
-import { BergamotApiClient } from "./BergamotApiClient";
+import { LanguageDetector } from "./lib/LanguageDetector";
 import { Runtime } from "webextension-polyfill-ts";
 import Port = Runtime.Port;
-const bergamotApiClient = new BergamotApiClient();
 
-export const contentScriptBergamotApiClientPortListener = (port: Port) => {
-  if (port.name !== "port-from-content-script-bergamot-api-client") {
+export const contentScriptLanguageDetectorProxyPortListener = (port: Port) => {
+  if (port.name !== "port-from-content-script-language-detector-proxy") {
     return;
   }
   port.onMessage.addListener(async function(m: {
-    texts: [];
+    str: string;
     requestId: string;
   }) {
-    // console.debug("Message from content-script-bergamot-api-client:", {m});
-    const { texts, requestId } = m;
-    const results = await bergamotApiClient.sendTranslationRequest(texts);
-    // console.log({ results });
+    // console.debug("Message from content-script-language-detector-proxy:", { m });
+    const { str, requestId } = m;
+    const results = await LanguageDetector.detectLanguage({ text: str });
+    // console.debug({ results });
     try {
       port.postMessage({
-        translationRequestResults: {
+        languageDetectorResults: {
           results,
           requestId,
         },
