--- extension/ts/background-scripts/background.js/lib/translation.constants.ts
@@ -1,6 +1,0 @@
-export const TRANSLATION_PREF_ENGINE = "browser.translation.engine";
-export const TRANSLATION_PREF_NEVER_FOR_LANGUAGES =
-  "browser.translation.neverForLanguages";
-export const TRANSLATION_PREF_SHOWUI = "browser.translation.ui.show";
-export const TRANSLATION_PREF_DETECT_LANG =
-  "browser.translation.detectLanguage";
