+++ extension/ts/background-scripts/background.js/state-management/Store.ts
