+++ extension/ts/background-scripts/background.js/state-management/localStorageWrapper.ts
