+++ extension/ts/background-scripts/background.js/state-management/mockLocalStorage.spec.ts
