+++ extension/ts/extension-ui/get-started.js/index.tsx
@@ -1,13 +1,9 @@
-import { initErrorReportingInContentScript } from "../../shared-resources//ErrorReporting";
-
+import { initErrorReportingInContentScript } from "../../shared-resources/ErrorReporting";
 import * as React from "react";
 import * as ReactDOM from "react-dom";
-
-import { ErrorBoundary } from "../../shared-resources//ErrorBoundary";
+import { ErrorBoundary } from "../../shared-resources/ErrorBoundary";
 import { GetStarted } from "./GetStarted";
-
 import "./index.css";
-
 import { DisplayError } from "./DisplayError";
 
 const init = async () => {
