+++ extension/ts/extension-ui/options-ui.js/index.tsx
@@ -1,9 +1,9 @@
-import { initErrorReportingInContentScript } from "../../shared-resources//ErrorReporting";
+import { initErrorReportingInContentScript } from "../../shared-resources/ErrorReporting";
 
 import * as React from "react";
 import * as ReactDOM from "react-dom";
 
-import { ErrorBoundary } from "../../shared-resources//ErrorBoundary";
+import { ErrorBoundary } from "../../shared-resources/ErrorBoundary";
 import { OptionsUi } from "./OptionsUi";
 import { DisplayError } from "./DisplayError";
