+++ extension/ts/shared-resources/ContentScriptBergamotApiClient.ts
@@ -1,7 +1,7 @@
 import { browser, Runtime } from "webextension-polyfill-ts";
+import Port = Runtime.Port;
 import { nanoid } from "nanoid";
 import { captureExceptionWithExtras } from "./ErrorReporting";
-import Port = Runtime.Port;
 
 export class ContentScriptBergamotApiClient {
   private backgroundContextPort: Port;
