+++ extension/ts/shared-resources/getCurrentTab.ts
