+++ extension/ts/shared-resources/state-management/subscribeToExtensionState.ts
@@ -11,10 +11,10 @@ import {
   setGlobalConfig,
 } from "mobx-keystone";
 import { browser, Runtime } from "webextension-polyfill-ts";
-import { captureExceptionWithExtras } from "./ErrorReporting";
-import { nanoid } from "nanoid";
 import Port = Runtime.Port;
-import { ExtensionState } from "./models/ExtensionState";
+import { captureExceptionWithExtras } from "../ErrorReporting";
+import { nanoid } from "nanoid";
+import { ExtensionState } from "../models/ExtensionState";
 
 // disable runtime data checking (we rely on TypeScript at compile time so that our model definitions can be cleaner)
 setGlobalConfig({
